+++ meta/recipes-lora/lora-concentrator/files/lora-concentrator-reset.sh
@@ -4,7 +4,6 @@
 
 CHIP=${LORA_CORE_RST_CHIP:-gpiochip0}
 PIN=${LORA_CORE_RST_PIN:-1}
-DELAY=${LORA_CORE_RST_DELAY_US:-100000}
+DELAY=${LORA_CORE_RST_DELAY_MS:-100}
 
-gpioset --mode=time --usec=${DELAY} ${CHIP} ${PIN}=1
-gpioset ${CHIP} ${PIN}=0
+gpioset --toggle "${DELAY}ms,0" --chip ${CHIP} ${PIN}=1