+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [LOR-632] Update Yocto to Scarthgap (5.0)
+- [LOR-687] Upgrade syslog-ng config to corresponding version
 
 ## [1.7.3] - 2025-06-26