+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [LOR-675] Upgrade the firewall to nftables
 - [LOR-687] Upgrade syslog-ng config to corresponding version
 - [LOR-700] Update Chrony to v4
+- [LOR-705] Forwarder logs are no more added to the "all" log
 
 ### Fixed